Break out IncrementVersion
