+++ .atomist/editors/BumpVersion.ts
@@ -18,6 +18,7 @@ import { Project } from "@atomist/rug/model/Project";
 import { Editor, Parameter, Tags } from "@atomist/rug/operations/Decorators";
 import { EditProject } from "@atomist/rug/operations/ProjectEditor";
 
+import { incrementVersion } from "./libbits/IncrementVersion";
 import { isRugArchive, NotRugArchiveError } from "./RugEditorsPredicates";
 
 /**
@@ -66,28 +67,3 @@ export class BumpVersion implements EditProject {
 }
 
 export const bumpVersion = new BumpVersion();
-
-export function incrementVersion(version: string, component: "major" | "minor" | "patch"): string {
-    const versionRegex = /^(\d+)\.(\d+)\.(\d+)([-.].*)?$/;
-    const versionMatch = versionRegex.exec(version);
-    if (versionMatch === null || versionMatch.length < 4) {
-        throw new Error(`version does not appear to be valid: ${version}`);
-    }
-
-    let major = parseInt(versionMatch[1], 10);
-    let minor = parseInt(versionMatch[2], 10);
-    let patch = parseInt(versionMatch[3], 10);
-    const rest = (versionMatch[4] != null) ? versionMatch[4] : "";
-
-    if (component === "major") {
-        major = major + 1;
-        minor = 0;
-        patch = 0;
-    } else if (component === "minor") {
-        minor = minor + 1;
-        patch = 0;
-    } else if (component === "patch") {
-        patch = patch + 1;
-    }
-    return `${major}.${minor}.${patch}${rest}`;
-}
